backend/requirements.txt
django==4.2.5
djangorestframework==3.14.0
django-filter==23.2
drf-yasg==1.21.7